Refactor backend utils.
