+++ litestar/dto/factory/_backends/utils/typing.py
@@ -1,0 +1,28 @@
+from __future__ import annotations
+
+from typing import Any, Iterable, TypeVar
+
+from typing_extensions import get_origin
+
+T = TypeVar("T")
+
+__all__ = ("build_annotation_for_backend",)
+
+
+def build_annotation_for_backend(annotation: Any, model: type[T]) -> type[T] | type[Iterable[T]]:
+    """A helper to re-build a generic outer type with new inner type.
+
+    Args:
+        annotation: The original annotation on the handler signature
+        model: The data container type
+
+    Returns:
+        Annotation with new inner type if applicable.
+    """
+    origin = get_origin(annotation)
+    if not origin:
+        return model
+    try:
+        return origin[model]  # type:ignore[no-any-return]
+    except TypeError:  # pragma: no cover
+        return annotation.copy_with((model,))  # type:ignore[no-any-return]
