+++ tests/contrib/sqlalchemy/test_dto_integration.py
@@ -10,7 +10,7 @@ from typing_extensions import Annotated
 from litestar import get, post
 from litestar.contrib.sqlalchemy.dto import SQLAlchemyDTO
 from litestar.dto.factory import DTOConfig
-from litestar.dto.factory._backends.utils import RenameStrategies
+from litestar.dto.factory._backends.utils.renaming import RenameStrategies
 from litestar.dto.factory.types import RenameStrategy
 from litestar.testing import create_test_client
